+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Server: Use Kadalu Volgen library for volfile generation
+
 ## [0.9.0] - 2022-11-21
 
 - Helm chart: fixed namespace management bug